execution: reorganize function operators
