+++ execution/function/noarg.go
@@ -1,0 +1,59 @@
+// Copyright (c) The Thanos Community Authors.
+// Licensed under the Apache License 2.0.
+
+package function
+
+import (
+	"context"
+	"fmt"
+
+	"github.com/prometheus/prometheus/model/labels"
+
+	"github.com/thanos-io/promql-engine/execution/model"
+	"github.com/thanos-io/promql-engine/parser"
+)
+
+type noArgFunctionOperator struct {
+	mint        int64
+	maxt        int64
+	step        int64
+	currentStep int64
+	stepsBatch  int
+	funcExpr    *parser.Call
+	call        functionCall
+	vectorPool  *model.VectorPool
+	series      []labels.Labels
+	sampleIDs   []uint64
+}
+
+func (o *noArgFunctionOperator) Explain() (me string, next []model.VectorOperator) {
+	return fmt.Sprintf("[*noArgFunctionOperator] %v()", o.funcExpr.Func.Name), []model.VectorOperator{}
+}
+
+func (o *noArgFunctionOperator) Series(_ context.Context) ([]labels.Labels, error) {
+	return o.series, nil
+}
+
+func (o *noArgFunctionOperator) GetPool() *model.VectorPool {
+	return o.vectorPool
+}
+
+func (o *noArgFunctionOperator) Next(_ context.Context) ([]model.StepVector, error) {
+	if o.currentStep > o.maxt {
+		return nil, nil
+	}
+	fa := functionArgs{}
+	ret := o.vectorPool.GetVectorBatch()
+	for i := 0; i < o.stepsBatch && o.currentStep <= o.maxt; i++ {
+		sv := o.vectorPool.GetStepVector(o.currentStep)
+		fa.StepTime = o.currentStep
+		result := o.call(fa)
+		sv.Samples = []float64{result.F}
+		sv.SampleIDs = o.sampleIDs
+
+		ret = append(ret, sv)
+		o.currentStep += o.step
+	}
+
+	return ret, nil
+}
