+++ execution/scan/matrix_selector.go
@@ -16,6 +16,7 @@ import (
 
 	"github.com/thanos-io/promql-engine/execution/function"
 	"github.com/thanos-io/promql-engine/execution/model"
+	"github.com/thanos-io/promql-engine/execution/parse"
 	engstore "github.com/thanos-io/promql-engine/execution/storage"
 	"github.com/thanos-io/promql-engine/parser"
 	"github.com/thanos-io/promql-engine/query"
@@ -24,7 +25,7 @@ import (
 type matrixScanner struct {
 	labels           labels.Labels
 	signature        uint64
-	previousSamples  []function.Sample
+	previousSamples  []sample
 	samples          *storage.BufferedSeriesIterator
 	metricAppearedTs *int64
 }
@@ -32,7 +33,7 @@ type matrixScanner struct {
 type matrixSelector struct {
 	funcExpr *parser.Call
 	storage  engstore.SeriesSelector
-	call     function.FunctionCall
+	call     functionCall
 	scanners []matrixScanner
 	series   []labels.Labels
 	once     sync.Once
@@ -54,17 +55,19 @@ type matrixSelector struct {
 	extLookbackDelta int64
 }
 
-// NewMatrixSelector creates operator which selects vector of series over time.
+// NewRangeVectorFunction creates operator which selects vector of series over time.
 func NewMatrixSelector(
 	pool *model.VectorPool,
 	selector engstore.SeriesSelector,
-	call function.FunctionCall,
 	funcExpr *parser.Call,
 	opts *query.Options,
 	selectRange, offset time.Duration,
 	shard, numShard int,
-) model.VectorOperator {
-	// TODO(fpetkovski): Add offset parameter.
+) (model.VectorOperator, error) {
+	call, ok := rangeVectorFuncs[funcExpr.Func.Name]
+	if !ok {
+		return nil, parse.UnknownFunctionError(funcExpr.Func)
+	}
 	return &matrixSelector{
 		storage:    selector,
 		call:       call,
@@ -84,7 +87,7 @@ func NewMatrixSelector(
 		numShards: numShard,
 
 		extLookbackDelta: opts.ExtLookbackDelta.Milliseconds(),
-	}
+	}, nil
 }
 
 func (o *matrixSelector) Explain() (me string, next []model.VectorOperator) {
@@ -140,10 +143,10 @@ func (o *matrixSelector) Next(ctx context.Context) ([]model.StepVector, error) {
 			maxt := seriesTs - o.offset
 			mint := maxt - o.selectRange
 
-			var rangeSamples []function.Sample
+			var rangeSamples []sample
 			var err error
 
-			if function.IsExtFunction(o.funcExpr.Func.Name) {
+			if parse.IsExtFunction(o.funcExpr.Func.Name) {
 				rangeSamples, err = selectExtPoints(series.samples, mint, maxt, o.scanners[i].previousSamples, o.extLookbackDelta, &o.scanners[i].metricAppearedTs)
 			} else {
 				rangeSamples, err = selectPoints(series.samples, mint, maxt, o.scanners[i].previousSamples)
@@ -157,7 +160,7 @@ func (o *matrixSelector) Next(ctx context.Context) ([]model.StepVector, error) {
 			// Also, allow operator to exist independently without being nested
 			// under parser.Call by implementing new data model.
 			// https://github.com/thanos-io/promql-engine/issues/39
-			result := o.call(function.FunctionArgs{
+			result := o.call(functionArgs{
 				Samples:          rangeSamples,
 				StepTime:         seriesTs,
 				SelectRange:      o.selectRange,
@@ -165,7 +168,7 @@ func (o *matrixSelector) Next(ctx context.Context) ([]model.StepVector, error) {
 				MetricAppearedTs: o.scanners[i].metricAppearedTs,
 			})
 
-			if result.T != function.InvalidSample.T {
+			if result.T != invalidSample.T {
 				vectors[currStep].T = result.T
 				if result.H != nil {
 					vectors[currStep].AppendHistogram(o.vectorPool, series.signature, result.H)
@@ -181,7 +184,7 @@ func (o *matrixSelector) Next(ctx context.Context) ([]model.StepVector, error) {
 			if stepRange > o.step {
 				stepRange = o.step
 			}
-			if !function.IsExtFunction(o.funcExpr.Func.Name) {
+			if !parse.IsExtFunction(o.funcExpr.Func.Name) {
 				series.samples.ReduceDelta(stepRange)
 			}
 
@@ -223,7 +226,7 @@ func (o *matrixSelector) loadSeries(ctx context.Context) error {
 
 			// If we are dealing with an extended range function we need to search further in the past for valid series.
 			var selectRange = o.selectRange
-			if function.IsExtFunction(o.funcExpr.Func.Name) {
+			if parse.IsExtFunction(o.funcExpr.Func.Name) {
 				selectRange += o.extLookbackDelta
 			}
 
@@ -248,7 +251,7 @@ func (o *matrixSelector) loadSeries(ctx context.Context) error {
 // into the [mint, maxt] range are retained; only points with later timestamps
 // are populated from the iterator.
 // TODO(fpetkovski): Add max samples limit.
-func selectPoints(it *storage.BufferedSeriesIterator, mint, maxt int64, out []function.Sample) ([]function.Sample, error) {
+func selectPoints(it *storage.BufferedSeriesIterator, mint, maxt int64, out []sample) ([]sample, error) {
 	if len(out) > 0 && out[len(out)-1].T >= mint {
 		// There is an overlap between previous and current ranges, retain common
 		// points. In most such cases:
@@ -285,7 +288,7 @@ loop:
 				continue loop
 			}
 			if t >= mint {
-				out = append(out, function.Sample{T: t, H: fh})
+				out = append(out, sample{T: t, H: fh})
 			}
 		case chunkenc.ValFloat:
 			t, v := buf.At()
@@ -294,7 +297,7 @@ loop:
 			}
 			// Values in the buffer are guaranteed to be smaller than maxt.
 			if t >= mint {
-				out = append(out, function.Sample{T: t, F: v})
+				out = append(out, sample{T: t, F: v})
 			}
 		}
 	}
@@ -304,12 +307,12 @@ loop:
 	case chunkenc.ValHistogram, chunkenc.ValFloatHistogram:
 		t, fh := it.AtFloatHistogram()
 		if t == maxt && !value.IsStaleNaN(fh.Sum) {
-			out = append(out, function.Sample{T: t, H: fh})
+			out = append(out, sample{T: t, H: fh})
 		}
 	case chunkenc.ValFloat:
 		t, v := it.At()
 		if t == maxt && !value.IsStaleNaN(v) {
-			out = append(out, function.Sample{T: t, F: v})
+			out = append(out, sample{T: t, F: v})
 		}
 	}
 
@@ -325,7 +328,7 @@ loop:
 // into the [mint, maxt] range are retained; only points with later timestamps
 // are populated from the iterator.
 // TODO(fpetkovski): Add max samples limit.
-func selectExtPoints(it *storage.BufferedSeriesIterator, mint, maxt int64, out []function.Sample, extLookbackDelta int64, metricAppearedTs **int64) ([]function.Sample, error) {
+func selectExtPoints(it *storage.BufferedSeriesIterator, mint, maxt int64, out []sample, extLookbackDelta int64, metricAppearedTs **int64) ([]sample, error) {
 	extMint := mint - extLookbackDelta
 
 	if len(out) > 0 && out[len(out)-1].T >= mint {
@@ -378,7 +381,7 @@ loop:
 				*metricAppearedTs = &t
 			}
 			if t >= mint {
-				out = append(out, function.Sample{T: t, H: fh})
+				out = append(out, sample{T: t, H: fh})
 			}
 		case chunkenc.ValFloat:
 			t, v := buf.At()
@@ -393,10 +396,10 @@ loop:
 			// exists at or before range start, add it and then keep replacing
 			// it with later points while not yet (strictly) inside the range.
 			if t >= mint || !appendedPointBeforeMint {
-				out = append(out, function.Sample{T: t, F: v})
+				out = append(out, sample{T: t, F: v})
 				appendedPointBeforeMint = true
 			} else {
-				out[len(out)-1] = function.Sample{T: t, F: v}
+				out[len(out)-1] = sample{T: t, F: v}
 			}
 
 		}
@@ -410,7 +413,7 @@ loop:
 			if *metricAppearedTs == nil {
 				*metricAppearedTs = &t
 			}
-			out = append(out, function.Sample{T: t, H: fh})
+			out = append(out, sample{T: t, H: fh})
 		}
 	case chunkenc.ValFloat:
 		t, v := it.At()
@@ -418,7 +421,7 @@ loop:
 			if *metricAppearedTs == nil {
 				*metricAppearedTs = &t
 			}
-			out = append(out, function.Sample{T: t, F: v})
+			out = append(out, sample{T: t, F: v})
 		}
 	}
